migrate to vim-plug
vundle has been working ok for me in the past. but for coc.nvim I need
to be able to clone a specific branch and vundle does not have this
functionality, and it hasn't moved for the last three years so I'm not
confident that it'll ever have it.

vim-plug is ... ridiculously fast! it runs all plugin installs in
parallel, so just confirming that all plugins are already installed task
about 0.3s. vundle was slower than this.

vim-plug's install method is a bit dumb though.. why isn't the upstream
repo formatted in a way that I can just git clone that somewhere and
then use it? aaaanyway