Change mapping for accepting completion
Mapping <CR> in I mode to some function is, pretty much as I was
expecting, very annoying. I'm moving it to ctrl+space but that means
that I don't have a mapping anymore for canceling the completion dialog.

I don't use the cancel function very much so it should be OK to have it
just not there while I think of a way to map it to something else.