finalize vimspector config and use only one install of puppet-editor-services
I've also included a default adapter config for puppet in my vimrc,
which removes the need to create `.vimspector.json` files everywhere.
I'll see if it's a bother, but since it set as default it should be
possible to override it per project.

With the MR I've sent for adding puppet support upstream to vimspector,
I no longer need to install puppet-editor-services manually. Hopefully
the MR will get merged..