Remove c-space mapping. Remove broken option completeopt
it works ok in vim but not in nvim. also there's already the builtin c-y
for accepting the selected completion.

completeopt used to look like it was set, but with a double quoted
sting. This actually was parsed as a comment so it wasn't being
considered. With coc.nvim the option doesn't change anything so might as
well just remove it from my vimrc