Replace auto-pairs with nvim-autopairs
the version of auto-pairs that I found is managed but they created
mappings on <C-P><something> that can't be deactivated with a global
variable. This interferes with the default <C-P> for selecting a
previous suggestion, which is very damn annoying -- it adds a 1s delay
every time I want to move back in the list. The nvim plugin achieves the
auto-insertion of matching pairs without the above annoying behavior.