Bring back vader.vim for syntax highlighting
vim-puppet is still using vader.vim for its tests and I need to be able
to edit them. having the module around means I'll have syntax
highlighting for the test files.

vader has been moving super slowly, but there was something merged in 5
months ago so it's not entirely dead yet. I've just poked the plugin's
author to get feedback on the project's status and suggest to maybe add
one or more ppl as maintainers to help out with merging PRs.