Change to black for autoformat and flake8 for linting
There's been a couple of cases where projects found problems with flake8
that I just didn't see. The linting from ruff seems to be pretty
minimal. So fuck that, let's move back to flake8 for linting.

the version of ruff in debian is too old for the autoformat used by
coc-pyright: it's trying to use an option that doesn't exist yet. and so
far I'm not getting the sense that folks in debian will make the move to
bump the version of ruff before trixie is released, which is pretty sad.
so if I use black instead at least I get some form of functional
autoformatting.