Fix colors again again
Why is this moving? It's been working OK for a good number of years.

I ended up figuring out that the colors were different because somehow
nvim now enables termguicolors by default even though it's documented to
be set to off by default. That means that it's using the `gui*` color
schemes.

I've reverted back to disabling termguicolors, but also adjusted the
gui colors just in case I ever want to switch to that. It's going to
look aweful in the actual gui, but I've never ever used that (after
figuring out that it looks like crap)

I've also added DiagnosticHint in there since whatever sets that (CoC?)
decided to revert to just plain "gray". That makes it impossible to spot
which type hints in python are manually added and which ones are
virtual.